change: apply pr fixes
