+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/edc/logic/util/EdcRequestBodyBuilder.java
@@ -313,21 +313,6 @@ public class EdcRequestBodyBuilder {
         return body;
     }
 
-    public JsonNode buildPartTypeInfoContractDefinitionForPartner(Partner partner) {
-        var body = getEdcContextObject();
-        body.put("@id", partner.getBpnl() + "_contractdefinition_for_PartTypeInfoAsset");
-        body.put("accessPolicyId", getBpnPolicyId(partner));
-        body.put("contractPolicyId", partner.getPolicyProfileVersion().CONTRACT_POLICY_ID);
-        var assetsSelector = MAPPER.createObjectNode();
-        body.set("assetsSelector", assetsSelector);
-        assetsSelector.put("@type", "CriterionDto");
-        assetsSelector.put("operandLeft", JsonLdConstants.EDC_NAMESPACE + "id");
-        assetsSelector.put("operator", "=");
-        assetsSelector.put("operandRight", getPartTypeInfoLegacyAssetId());
-        assetsSelector.put("operandRight", getPartTypeInfoAssetId());
-        return body;
-    }
-
     /**
      * This method helps to ensure that the buildContractDefinitionWithBpnRestrictedPolicy uses the
      * same policy-id as the one that is created with the buildContractDefinitionWithBpnRestrictedPolicy
@@ -528,74 +513,8 @@ public class EdcRequestBodyBuilder {
         return dataAddress;
     }
 
-    public JsonNode buildPartTypeInfoLegacySubmodelRegistrationBody() {
-        var body = getAssetRegistrationContext();
-        body.put("@id", getPartTypeInfoLegacyAssetId());
-        var propertiesObject = MAPPER.createObjectNode();
-        body.set("properties", propertiesObject);
-        var dctTypeObject = MAPPER.createObjectNode();
-        propertiesObject.set("dct:type", dctTypeObject);
-        dctTypeObject.put("@id", JsonLdConstants.CX_TAXO_NAMESPACE + "Submodel");
-        propertiesObject.put("cx-common:version", "3.0");
-        var semanticIdObject = MAPPER.createObjectNode();
-        propertiesObject.set("aas-semantics:semanticId", semanticIdObject);
-        semanticIdObject.put("@id", "urn:samm:io.catenax.part_type_information:1.0.0#PartTypeInformation");
-        var dataAddress = MAPPER.createObjectNode();
-        String url = variablesService.getParttypeInformationLegacyServerendpoint();
-        if (!url.endsWith("/")) {
-            url += "/";
-        }
-        dataAddress.put("@type", "DataAddress");
-        dataAddress.put("proxyPath", "true");
-        dataAddress.put("proxyQueryParams", "false");
-        dataAddress.put("proxyMethod", "false");
-        dataAddress.put("type", "HttpData");
-        dataAddress.put("baseUrl", url);
-        dataAddress.put("authKey", "x-api-key");
-        dataAddress.put("authCode", variablesService.getApiKey());
-        body.set("dataAddress", dataAddress);
-        return body;
-    }
-
-    public JsonNode buildPartTypeInfoSubmodelRegistrationBody() {
-        var body = getAssetRegistrationContext();
-        body.put("@id", getPartTypeInfoAssetId());
-        var propertiesObject = MAPPER.createObjectNode();
-        body.set("properties", propertiesObject);
-        var dctTypeObject = MAPPER.createObjectNode();
-        propertiesObject.set("dct:type", dctTypeObject);
-        dctTypeObject.put("@id", JsonLdConstants.CX_TAXO_NAMESPACE + "Submodel");
-        propertiesObject.put("cx-common:version", "3.0");
-        var semanticIdObject = MAPPER.createObjectNode();
-        propertiesObject.set("aas-semantics:semanticId", semanticIdObject);
-        semanticIdObject.put("@id", "urn:samm:io.catenax.part_type_information:2.0.0#PartTypeInformation");
-        var dataAddress = MAPPER.createObjectNode();
-        String url = variablesService.getParttypeInformationServerendpoint();
-        if (!url.endsWith("/")) {
-            url += "/";
-        }
-        dataAddress.put("@type", "DataAddress");
-        dataAddress.put("proxyPath", "true");
-        dataAddress.put("proxyQueryParams", "false");
-        dataAddress.put("proxyMethod", "false");
-        dataAddress.put("type", "HttpData");
-        dataAddress.put("baseUrl", url);
-        dataAddress.put("authKey", "x-api-key");
-        dataAddress.put("authCode", variablesService.getApiKey());
-        body.set("dataAddress", dataAddress);
-        return body;
-    }
-
     private String getDtrAssetId() {
         return "DigitalTwinRegistryId@" + variablesService.getOwnBpnl();
-    }
-
-    private String getPartTypeInfoLegacyAssetId() {
-        return variablesService.getPartTypeLegacySubmodelApiAssetId();
-    }
-
-    private String getPartTypeInfoAssetId() {
-        return variablesService.getPartTypeSubmodelApiAssetId();
     }
 
     private ObjectNode getAssetRegistrationContext() {
