+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/logic/dto/parttypeinformation/v1/PartSitesInformationAsPlanned.java
@@ -30,6 +30,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
+import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.parttypeinformation.FunctionEnum;
 
 import java.util.Objects;
