+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/masterdata/logic/dto/parttypeinformation/v2/PartSitesInformationAsPlanned.java
@@ -27,7 +27,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 import org.eclipse.tractusx.puris.backend.common.util.PatternStore;
-import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.parttypeinformation.v1.FunctionEnum;
+import org.eclipse.tractusx.puris.backend.masterdata.logic.dto.parttypeinformation.FunctionEnum;
 
 import java.util.Objects;
